Create TodoList component
